added filter configuration for domains
work on the TODOs to help improve the overall functionality of the module. it would be nice not to have to run the module over and over until it stops grabbing imgur album webpages. i'd also like to break it up into small modules (i.e. imgur images and cli).